Finish Implementing the very first version of SVE_Vectorizer

+++ Transformation Pass/Vectorizer Pass/test/test.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-void foo(float *__restrict__ a, float *__restrict__ b, float *__restrict__ c, int n) {
+void foo(int *__restrict__ a, int *__restrict__ b, int *__restrict__ c, int n) {
 
 //    #pragma clang loop unroll_count(4)
     for (int i = 0; i < n; ++i) {
@@ -12,14 +12,14 @@ void foo(float *__restrict__ a, float *__restrict__ b, float *__restrict__ c, in
 
 int main() {
     int n = 8;
-    float a[] = {1, -1, 2, -2, 3, -3, 4, -4};
-    float b[] = {2, 2, 2, 2, 2, 2, 2, 2};
-    float c[] = {0, 0, 0, 0, 0, 0, 0, 0};
+    int a[] = {1, -1, 2, -2, 3, -3, 4, -4};
+    int b[] = {2, 2, 2, 2, 2, 2, 2, 2};
+    int c[] = {0, 0, 0, 0, 0, 0, 0, 0};
 
     foo(a, b, c, n);
 
     for (int i = 0; i < n; ++i) {   // 0, -2, 0, -4, 0, -6, 0, -8
-        printf("%f ", c[i]);
+        printf("%d ", c[i]);
     }
     printf("\n");
     return 0;
